endpoint 优化
