--- alias.go
@@ -1,10 +1,0 @@
-package luchen
-
-import "github.com/go-kit/kit/endpoint"
-
-type (
-	// Endpoint alias for endpoint.Endpoint
-	Endpoint = endpoint.Endpoint
-	// Middleware alias for endpoint.Endpoint
-	Middleware = endpoint.Middleware
-)
